+++ CHANGES.rst
@@ -253,6 +253,9 @@ Bug Fixes
   - Avoid importing matplotlib.pyplot when importing
     astropy.visualization.wcsaxes. [#5680, #5684]
 
+  - Ignore Numpy warnings that happen in coordinate transforms in WCSAxes.
+    [#5792]
+
   - Fix compatibility issues between WCSAxes and Matplotlib 2.x. [#5786]
 
 - ``astropy.vo``